+++ packages/qiskit-sdk/bin/cmds/cloudJobs.js
@@ -11,25 +11,29 @@
 const qiskit = require('../..');
 const logger = require('../lib/logger');
 
-exports.command = 'qe-backs [onlySims]';
+exports.command = 'cloud-jobs [limit] [offset]';
 
-exports.aliases = ['qbs'];
+exports.aliases = ['cjs'];
 
-exports.desc = 'Get latest calibration stats for a backend';
+exports.desc = 'Get all your jobs. Ordered by creation date';
 
 exports.builder = {
-  onlySims: {
+  limit: {
     desc: ' To get only info of the simulators',
-    type: 'boolean',
-    default: false,
+    type: 'number',
+    default: 50,
+  },
+  offset: {
+    desc: ' To get only info of the simulators',
+    type: 'number',
   },
 };
 
 exports.handler = argv => {
   logger.title(qiskit.version);
 
-  global.qiskit.qe
-    .backends(argv.onlySims)
+  global.qiskit.cloud
+    .jobs(argv.limit, argv.offset)
     .then(res => {
       logger.resultHead();
       logger.chunks(res);